@autowire Annotation.
this is the case where you can see there are two beans of Point class i.e pointA and pointB.
if we run the application there will be error as we autowired the point in circle bean. hence it will confuse whether to take pointA or pointB, and hence the second priority it give is to see the id name, hence when swapping the name pointA to center, the autowire works perfectly.