@autowire Annotation.
this is the case where you can see there are two beans of Point class i.e pointA and pointB.
if we run the application there will be error as we autowired the point in circle bean. hence it will confuse whether to take pointA or pointB, and hence the second priority also dont work as the id name is not center. so we use qualifier.

Make sure you give namespace at the top, otherwise it gonna give error that there is no qualifier tag.